- Convert some long-obsolete xlib calls into something newer and more
  general.

- Encpsulate some inline code into funcs for better readability, variable
  scope limit, etc.  Little inter-file moving as part of that.

- Cleanup some unnecessary casts.

- Misc minor cleanup and commenting.

+++ ewmh.c
@@ -988,10 +988,10 @@ static void EwmhHandle_NET_WM_ICONNotify(XPropertyEvent *event,
 	x = GetIconOffset(twm_win->icon);
 	twm_win->icon->bm_w =
 	        XCreateWindow(dpy, twm_win->icon->w, x, 0,
-	                      (unsigned int) twm_win->icon->width,
-	                      (unsigned int) twm_win->icon->height,
-	                      (unsigned int) 0, Scr->d_depth,
-	                      (unsigned int) CopyFromParent, Scr->d_visual,
+	                      twm_win->icon->width,
+	                      twm_win->icon->height,
+	                      0, Scr->d_depth,
+	                      CopyFromParent, Scr->d_visual,
 	                      valuemask, &attributes);
 
 	if(image->mask) {
